scripts: drop resource manager specific subdirs
