+++ scripts/python/scrjob/scr_inspect.py
@@ -77,7 +77,10 @@ def scr_inspect(jobnodes=None, up=None, down=None, cntldir=None, scr_env=None):
     errfile = os.path.join(scr_dir, 'scr_inspect.pdsh.e.' + jobid)
 
     # run scr_inspect_cache via pdsh / clustershell
-    argv = [os.path.join(libexecdir, 'scr_inspect_cache'), os.path.join(cntldir, 'filemap.scrinfo')]
+    argv = [
+        os.path.join(libexecdir, 'scr_inspect_cache'),
+        os.path.join(cntldir, 'filemap.scrinfo')
+    ]
     out = scr_env.launcher.parallel_exec(argv=argv, runnodes=upnodes)[0]
     try:
         with open(outfile, 'w') as f:
@@ -132,7 +135,8 @@ def scr_inspect(jobnodes=None, up=None, down=None, cntldir=None, scr_env=None):
 
     except Exception as e:
         print(e)
-        raise RuntimeError('Reading and processing output file \"' + output + '\"')
+        raise RuntimeError('Reading and processing output file \"' + output +
+                           '\"')
 
     # starting with the most recent dataset, check whether we have (or may be able to recover) all files
     possible_dsets = []
@@ -180,7 +184,9 @@ def scr_inspect(jobnodes=None, up=None, down=None, cntldir=None, scr_env=None):
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(
-        epilog='The jobid and job node set must be able to be obtained from the environment.')
+        epilog=
+        'The jobid and job node set must be able to be obtained from the environment.'
+    )
     parser.add_argument('-j',
                         '--jobset',
                         default=None,
@@ -212,9 +218,9 @@ if __name__ == '__main__':
 
     try:
         dsets = scr_inspect(jobnodes=args.jobset,
-                          up=args.up,
-                          down=args.down,
-                          cntldir=args['from'])
+                            up=args.up,
+                            down=args.down,
+                            cntldir=args['from'])
         print(' '.join(dsets))
     except Exception as e:
         print('scr_inspect: ERROR: ' + str(e))
